Removed libraries in project, now they're replaced by find_package from CMake.

+++ renderer/vertex_array/VertexArray.cpp
@@ -40,7 +40,7 @@ VertexArray& VertexArray::setVBOData(TypeVertices vertices, numeralTypes::enumIn
 VertexArray& VertexArray::setIBOData(TypeIndices indexes, numeralTypes::enumInt usage) {
 	std::vector<unsigned short> indicesVector;
 	for (std::array<unsigned short, 3>& face : indexes) {
-		for (float element : face) {
+		for (unsigned short element : face) {
 			indicesVector.push_back(element);
 		}
 	}
